finish merge with FTC HQ release
Former-commit-id: 75c90bab1ba8beb4eaaf86d4a8f45e9dfa22e688
Former-commit-id: 119ecb731e1365d3d5ee54513366845ea0584657

+++ SwerveRoboticsLibrary/src/main/java/org/swerverobotics/library/internal/ThreadSafePWMOutput.java
@@ -35,7 +35,7 @@ public class ThreadSafePWMOutput extends PWMOutput
         super.setPulseWidthOutputTime(time);
         }
 
-    @Override public synchronized double getPulseWidthOutputTime() 
+    @Override public synchronized int getPulseWidthOutputTime()
         {
         return super.getPulseWidthOutputTime();
         }
@@ -45,7 +45,7 @@ public class ThreadSafePWMOutput extends PWMOutput
         super.setPulseWidthPeriod(period);
         }
 
-    @Override public synchronized double getPulseWidthPeriod() 
+    @Override public synchronized int getPulseWidthPeriod()
         {
         return super.getPulseWidthPeriod();
         }
